+++ tests/qchem/test_tapering.py
@@ -22,7 +22,6 @@ import scipy
 
 import pennylane as qml
 from pennylane import numpy as np
-from pennylane.math.utils import binary_finite_reduced_row_echelon
 from pennylane.pauli import pauli_sentence
 from pennylane.qchem.tapering import (
     _kernel,
@@ -33,97 +32,6 @@ from pennylane.qchem.tapering import (
     taper_hf,
     taper_operation,
 )
-
-
-@pytest.mark.parametrize(
-    ("binary_matrix", "result"),
-    [
-        (
-            np.array(
-                [
-                    [0, 0, 0, 0, 0, 0, 0, 0],
-                    [1, 0, 0, 0, 0, 0, 0, 0],
-                    [0, 1, 0, 0, 0, 0, 0, 0],
-                    [0, 0, 1, 0, 0, 0, 0, 0],
-                    [0, 0, 0, 1, 0, 0, 0, 0],
-                    [1, 1, 0, 0, 0, 0, 0, 0],
-                    [1, 0, 0, 1, 1, 1, 1, 1],
-                    [1, 1, 0, 0, 1, 1, 1, 1],
-                    [0, 0, 1, 1, 1, 1, 1, 1],
-                    [0, 1, 1, 0, 1, 1, 1, 1],
-                    [1, 0, 1, 0, 0, 0, 0, 0],
-                    [1, 0, 0, 1, 0, 0, 0, 0],
-                    [0, 1, 1, 0, 0, 0, 0, 0],
-                    [0, 1, 0, 1, 0, 0, 0, 0],
-                    [0, 0, 1, 1, 0, 0, 0, 0],
-                ]
-            ),
-            np.array(
-                [
-                    [1, 0, 0, 0, 0, 0, 0, 0],
-                    [0, 1, 0, 0, 0, 0, 0, 0],
-                    [0, 0, 1, 0, 0, 0, 0, 0],
-                    [0, 0, 0, 1, 0, 0, 0, 0],
-                    [0, 0, 0, 0, 1, 1, 1, 1],
-                    [0, 0, 0, 0, 0, 0, 0, 0],
-                    [0, 0, 0, 0, 0, 0, 0, 0],
-                    [0, 0, 0, 0, 0, 0, 0, 0],
-                    [0, 0, 0, 0, 0, 0, 0, 0],
-                    [0, 0, 0, 0, 0, 0, 0, 0],
-                    [0, 0, 0, 0, 0, 0, 0, 0],
-                    [0, 0, 0, 0, 0, 0, 0, 0],
-                    [0, 0, 0, 0, 0, 0, 0, 0],
-                    [0, 0, 0, 0, 0, 0, 0, 0],
-                    [0, 0, 0, 0, 0, 0, 0, 0],
-                ]
-            ),
-        ),
-    ],
-)
-def test_reduced_row_echelon(binary_matrix, result):
-    r"""Test that _reduced_row_echelon returns the correct result."""
-
-    # build row echelon form of the matrix
-    shape = binary_matrix.shape
-    for irow in range(shape[0]):
-        pivot_index = 0
-        if np.count_nonzero(binary_matrix[irow, :]):
-            pivot_index = np.nonzero(binary_matrix[irow, :])[0][0]
-
-        for jrow in range(shape[0]):
-            if jrow != irow and binary_matrix[jrow, pivot_index]:
-                binary_matrix[jrow, :] = (binary_matrix[jrow, :] + binary_matrix[irow, :]) % 2
-
-    indices = [
-        irow
-        for irow in range(shape[0] - 1)
-        if np.array_equal(binary_matrix[irow, :], np.zeros(shape[1]))
-    ]
-
-    temp_row_echelon_matrix = binary_matrix.copy()
-    for row in indices[::-1]:
-        temp_row_echelon_matrix = np.delete(temp_row_echelon_matrix, row, axis=0)
-
-    row_echelon_matrix = np.zeros(shape, dtype=int)
-    row_echelon_matrix[: shape[0] - len(indices), :] = temp_row_echelon_matrix
-
-    # build reduced row echelon form of the matrix from row echelon form
-    for idx in range(len(row_echelon_matrix))[:0:-1]:
-        nonzeros = np.nonzero(row_echelon_matrix[idx])[0]
-        if len(nonzeros) > 0:
-            redrow = (row_echelon_matrix[idx, :] % 2).reshape(1, -1)
-            coeffs = (
-                (-row_echelon_matrix[:idx, nonzeros[0]] / row_echelon_matrix[idx, nonzeros[0]]) % 2
-            ).reshape(1, -1)
-            row_echelon_matrix[:idx, :] = (
-                row_echelon_matrix[:idx, :] + (coeffs.T * redrow) % 2
-            ) % 2
-
-    # get reduced row echelon form from the _reduced_row_echelon function
-    rref_bin_mat = binary_finite_reduced_row_echelon(binary_matrix)
-
-    assert (rref_bin_mat == row_echelon_matrix).all()
-    assert (rref_bin_mat == result).all()
 
 
 @pytest.mark.parametrize(
